Add LxExec and LxExt
LxExt is an ld.preload extension for the init daemon, which creates a
UNIX Domain socket /tmp/lxexec-socket and waits for connections. Once
data is received, it passes it straight through the lxserver ADSS server
on the Win32 side. This removes the need to have the Registry key set
for ADSS Bus Access.
LsExec is a client for LxExt. It connects to the UNIX Domain Socket and
sends the input command line to it. This avoids needing to run as root,
such as lxclient before.