Implement LxDrv
LxDrv interacts with the new RS2 Device Minor APIs to register a new
0xBAD Minor device under the Misc. Device Major with the name
/dev/lxdrv. It can then be opened and interacted with by currently
sending it an IOCTL 0xBEEF and a 4-byte buffer, which will return 42.

NOTE: Never unload the driver with a running LX instance, and never
attempt creating new LX instances with the driver unloaded -- either
re-load the driver or reboot the machine.